backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.31
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.5
jinja2==3.1.5
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1